refactor: add small refactors
