+++ apps/web-neutron/src/components/nav/components/menu_items/utils.tsx
@@ -32,11 +32,6 @@ export const getMenuItems = () => {
       url: TRANSACTIONS,
       icon: <TransactionIcon {...iconProps} />,
     },
-    // {
-    //   key: 'proposals',
-    //   url: PROPOSALS,
-    //   icon: <ProposalsIcon {...iconProps} />,
-    // },
     {
       key: 'params',
       url: PARAMS,
